+++ apps/modernization-ui/src/apps/report/management/configuration/constants.ts
@@ -1,0 +1,2 @@
+
+export const SIZING = 'medium';
