+++ apps/modernization-ui/src/apps/report/management/configuration/utils.ts
@@ -1,0 +1,12 @@
+import { Selectable } from 'options';
+
+export const addLabelToName = ({ value, name, label }: Selectable) => ({
+    value,
+    label,
+    name: !label || label === name ? name : `${name} (${label})`,
+});
+
+export interface EnumSelectable<T> {
+    value: T;
+    name: string;
+}
